+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Support for yaml merge
 * MwsNccBinding databricks resource
 * MwsNetworkConnectivityConfig databricks resource
+* Support for `BINARYFILE` (PDF) format in `FileDataSource` with Spark 
 ### Fixed
 * DLT Debug mode when source is streaming and node is not
 * DataFrame type propagation when models are used as inputs to other models